ir_cdk_stacks/ir_cdk_stacks_stack.py
from aws_cdk import core


class IrCdkStacksStack(core.Stack):
    def __init__(self, scope: core.Construct, id: str, **kwargs) -> None:
        super().__init__(scope, id, **kwargs)

        # TODO: install global WAF here
