requirements.txt
appdirs==1.4.3
attrs==19.3.0
aws-cdk.assets==1.32.0
aws-cdk.aws-apigateway==1.32.0
aws-cdk.aws-applicationautoscaling==1.32.0
aws-cdk.aws-autoscaling==1.32.0
aws-cdk.aws-autoscaling-common==1.32.0
aws-cdk.aws-autoscaling-hooktargets==1.32.0
aws-cdk.aws-batch==1.32.0
aws-cdk.aws-certificatemanager==1.32.0
aws-cdk.aws-cloudformation==1.32.0
aws-cdk.aws-cloudfront==1.32.0
aws-cdk.aws-cloudwatch==1.32.0
aws-cdk.aws-cloudwatch-actions==1.32.0
aws-cdk.aws-dax==1.32.0
aws-cdk.aws-dynamodb==1.32.0
aws-cdk.aws-ec2==1.32.0
aws-cdk.aws-ecr==1.32.0
aws-cdk.aws-ecr-assets==1.32.0
aws-cdk.aws-ecs==1.32.0
aws-cdk.aws-elasticloadbalancing==1.32.0
aws-cdk.aws-elasticloadbalancingv2==1.32.0
aws-cdk.aws-events==1.32.0
aws-cdk.aws-glue==1.32.0
aws-cdk.aws-iam==1.32.0
aws-cdk.aws-kinesis==1.32.0
aws-cdk.aws-kms==1.32.0
aws-cdk.aws-lambda==1.32.0
aws-cdk.aws-lambda-event-sources==1.32.0
aws-cdk.aws-logs==1.32.0
aws-cdk.aws-logs-destinations==1.32.0
aws-cdk.aws-rds==1.32.0
aws-cdk.aws-route53==1.32.0
aws-cdk.aws-route53-targets==1.32.0
aws-cdk.aws-s3==1.32.0
aws-cdk.aws-s3-assets==1.32.0
aws-cdk.aws-s3-notifications==1.32.0
aws-cdk.aws-sam==1.32.0
aws-cdk.aws-secretsmanager==1.32.0
aws-cdk.aws-servicediscovery==1.32.0
aws-cdk.aws-sns==1.32.0
aws-cdk.aws-sns-subscriptions==1.32.0
aws-cdk.aws-sqs==1.32.0
aws-cdk.aws-ssm==1.32.0
aws-cdk.aws-stepfunctions==1.32.0
aws-cdk.aws-stepfunctions-tasks==1.32.0
aws-cdk.aws-wafv2==1.32.0
aws-cdk.core==1.32.0
aws-cdk.custom-resources==1.32.0
aws-cdk.cx-api==1.32.0
aws-cdk.region-info==1.32.0
black==19.10b0
cattrs==1.0.0
click==7.1.1
constructs==2.0.1
entrypoints==0.3
flake8==3.7.9
flake8-annotations==2.1.0
importlib-metadata==1.6.0
-e .
jsii==1.1.0
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pathspec==0.7.0
pluggy==0.13.1
pprint==0.1
prompt-toolkit==1.0.14
psycopg2==2.8.5
psycopg2-binary==2.8.5
publication==0.0.3
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
PyInquirer==1.0.3
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
regex==2020.4.4
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
wcwidth==0.1.9
zipp==3.1.0
